+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - ts-ebml plugin: fix Safari issue with ts-ebml (#595, #581)
 - Bump required version for videojs-wavesurfer (3.8.0 or newer)
+- Add `countdown` option for displaying 3-2-1 countdown before the recording (#604)
 
 
 ## 4.5.0 - 2021/06/14